requirements.txt
fastapi>=0.104.0
sqlalchemy>=2.0.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
uvicorn>=0.24.0
python-multipart>=0.0.6

